Do not copy non existing

+++ report-jtreg-list/src/main/java/io/jenkins/plugins/report/jtreg/main/recreate/ReportSummaryUtil.java
@@ -187,7 +187,9 @@ public class ReportSummaryUtil {
         Files.createDirectories(outDir);
         for (Path file : allFiles) {
             Path targetPath = outDir.resolve(file.getFileName());
-            Files.copy(file, targetPath, StandardCopyOption.REPLACE_EXISTING);
+            if (Files.exists(file)) {
+                Files.copy(file, targetPath, StandardCopyOption.REPLACE_EXISTING);
+            }
         }
     }
